==> FormerLabMembers/5k study R codes/8. EN model of CEL.Rmd <==
---
title: "EN models of CEL"
output:
  html_document:
    df_print: paged
  pdf_document: default
  html_notebook: default
---

THIS SCRIPT generates EN model of CEL in
training dataset contains all samples
validation dataset contains first LP only 

The model considers all high SNR markers, excluding NEFL and NEFH

###################

Set the environment 

```{r}

remove(list = ls())

library(readr, quietly = TRUE)
library(readxl, quietly = TRUE)
library(dplyr, quietly = TRUE)
library(tidyr, quietly = TRUE)
library(lubridate, quietly = TRUE)
library(purrr, quietly = TRUE)
library(bindrcpp, quietly = TRUE)
library(data.table, quietly = TRUE)
library(stringr, quietly = TRUE)
library(ggplot2, quietly = TRUE)
library(tidyr, quietly = TRUE)
library(corrr, quietly = TRUE)
library(Hmisc, quietly = TRUE)
library(foreach, quietly = TRUE)
library(diptest, quietly = TRUE)
library(glmnet)
library(broom)
library(plyr)
library(epiR)
library(Metrics)
library(ggpubr)

```


load functions

```{r}
'%notin%' <- function(x,y)!('%in%'(x,y))


tsapply <- function(...){t(sapply(...))}

cor_est <- function(x,y){
  obj <- cor.test(x,y,method = "spearman")
  val <- obj$estimate
  return(round(as.numeric(val),digits = 3))
}
cor_pval <- function(x,y){
  obj <- cor.test(x,y,method = "spearman")
  pval <- obj$p.value
  return(as.numeric(pval))
}


corr_fun <- function(dat,outcomes,high_markers,.){
    dat_est <- sapply(outcomes,function(ychar){
    sapply(high_markers,function(xchar){
      cor_est(dat[[xchar]],dat[[ychar]])
    })
  })
  
 
  dat_pval <- sapply(outcomes,function(ychar){sapply(high_markers,function(xchar){
    cor_pval(dat[[xchar]],dat[[ychar]])})})
  
  
  dat_pval <- as_tibble(dat_pval) %>% 
    mutate_all(p.adjust,method="fdr") %>%
    mutate(marker = high_markers,type="p-value") %>%  
    dplyr::select(marker,type,everything()) 
  
  dat_est <- as_tibble(dat_est) %>% 
    mutate(marker = high_markers,type="spearman") %>% 
    dplyr::select(marker,type,everything())
  
  dat_results <- bind_rows(dat_est,dat_pval) # stack em up
  dat_results <- dat_results %>% 
    gather(key,value,outcomes) 
 
  
  dat_results <- dat_results %>% 
    mutate(key = factor(key,levels=unique(dat_results$key))) %>% 
    mutate(type=factor(type,levels=c("spearman","p-value"))) %>% 
    arrange(marker,key,type) %>%    
    mutate(key = paste(as.character(key),type)) %>% 
    dplyr::select(-type) 
  
  dat_results <- dat_results %>% 
    mutate(key = factor(key,levels=unique(dat_results$key))) %>% 
    spread(key,value) 
  
  
  
  return(dat_results)
}

corr_fun_nofdr <- function(dat,outcomes,high_markers,.){
    dat_est <- sapply(outcomes,function(ychar){
    sapply(high_markers,function(xchar){
      cor_est(dat[[xchar]],dat[[ychar]])
    })
  })
  
 
  dat_pval <- sapply(outcomes,function(ychar){sapply(high_markers,function(xchar){
    cor_pval(dat[[xchar]],dat[[ychar]])})})
  
  
  dat_pval <- as_tibble(dat_pval) %>% 
    # mutate_all(p.adjust,method="fdr") %>%
    mutate(marker = high_markers,type="p-value") %>%  
    dplyr::select(marker,type,everything()) 
  
  dat_est <- as_tibble(dat_est) %>% 
    mutate(marker = high_markers,type="spearman") %>% 
    dplyr::select(marker,type,everything())
  
  dat_results <- bind_rows(dat_est,dat_pval) # stack em up
  dat_results <- dat_results %>% 
    gather(key,value,outcomes) 
 
  
  dat_results <- dat_results %>% 
    mutate(key = factor(key,levels=unique(dat_results$key))) %>% 
    mutate(type=factor(type,levels=c("spearman","p-value"))) %>% 
    arrange(marker,key,type) %>%    
    mutate(key = paste(as.character(key),type)) %>% 
    dplyr::select(-type) 
  
  dat_results <- dat_results %>% 
    mutate(key = factor(key,levels=unique(dat_results$key))) %>% 
    spread(key,value) 
  
  
  
  return(dat_results)
}



```

Load raw data

```{r}


# outcomes
demo_5k <- read_csv("./input/outcomes_5k.csv")

# translation file
transl_5k <- read_excel("./input/translation file 5k new.xlsx")

# adj & scaled somamers
soma_scaled <- fread("./input/soma_highSNR_adj_scaled.csv")

# high-signaling somamers
high_markers <- names(soma_scaled)[-1]


markers_5k <- transl_5k$SomaID
seqid_5k <- transl_5k$seqid


#merge outcomes with soma data
data_all <- merge(demo_5k,soma_scaled,by="sampleid",all.y = TRUE)


# plot histogram of trainnig and validation CEL distribution
data_train <- data_all %>% 
  filter(cohort_split=="training") %>% 
  filter(!is.na(cel)) 
# %>% 
#   filter(nice_therapy=="Untreated")

data_val <- data_all %>% 
  filter(cohort_split=="validation") %>% 
  filter(!is.na(cel))
# %>% 
#   filter(nice_therapy=="Untreated")
```


#####################################################################
####   EN models of CEL using all somamers
####   WITHOUT NEFL and NEFH - excluding pathogens somamers
####   
#####################################################################

```{r}

# identify NEFL and NEFH

nef <- c(transl_5k$seqid[which(transl_5k$Gene=="NEFL")],
         transl_5k$seqid[which(transl_5k$Gene=="NEFH")])


# make a matrix for saving predictions
data_out <- as.matrix(data_all[,high_markers])


marker <- "cel"
add <- "5k_train_all_val_first_lp_GLMNET_STANDARDIZED_all_soma_withoutNEF_noPathogen"
addd <- "all_soma_withoutNEF_noPathogen"
new_markers <- noPath_markers
#remove NEFL and NEFH
new_markers <- new_markers[-which(new_markers %in% nef)]

soma_ms_en <- data %>% 
  select(sampleid,cohort_split,all_of(marker),all_of(new_markers))

cv_list_soma_ms <- model_coefs_soma_ms <- list()

lm_models <- list()

stat_results <- tibble::enframe(marker,name = NULL,value = "marker")

stats_all_train <- tibble() 
stats_all_val <- tibble() 

  

  j <- "cel"
  
  #prepare training dataset
    dat <- data_train
  x_mean <- sapply(dat[,new_markers],mean)
  x_sd <- sapply(dat[,new_markers],sd)
  x <- as.matrix(dat[,new_markers]) 
  y <- dat[[j]]
  
  #prepare validation dataset
  dat_val <- data_val %>% 
    group_by(patientcode) %>% 
    arrange(lpdate) %>% 
    slice(1) %>% 
    ungroup()
  
  x_val <- as.matrix(dat_val[,new_markers])
  y_val <- dat_val[[j]]
 
  
  dat_all <- rbind(dat,dat_val)
  x_all <- as.matrix(dat_all[,new_markers])
  y_all <- dat_all[[j]]
 
  
  set.seed(1234)
  k <- 10
  foldid <- sample(rep(1:k,length.out=nrow(x)))
  # alpha_seq <- c(0,.1,.3,.5,.7,.9,1)
  # alpha_seq <- seq(0,1,length=20)
  alpha_seq <- seq(0,1,length=20)[-1]
  
  cv_error <- numeric(length=length(alpha_seq)) 
  for(i in 1:length(alpha_seq)){
    cv_out <- cv.glmnet(x,y,alpha=alpha_seq[i],standardize=TRUE,intercept=FALSE,
                        foldid = foldid) #alpha = 1 - lasso, alpha = 0 - ridge, between - elastic nets 
    cv_error[i] <- with(cv_out,cvm[which(lambda==lambda.1se)])
  }
  alpha_best <- alpha_seq[which.min(cv_error)]
  cv_out <- cv.glmnet(x,y,alpha=alpha_best,standardize=TRUE,intercept=TRUE, foldid = foldid) #alpha = 1 - lasso, alpha = 0 - ridge, between - elastic nets 
  cv_list_soma_ms[[j]] <- cv_out
  with(cv_out,cvm[which(lambda==lambda.1se)])
  
  
  # use previous tidy code to get model coefficients and save them using the paste() stuff
  glmnet_out <- tidy(cv_out$glmnet.fit) # glmnet_out has all lambdas considered
  glmnet_fit_best <- filter(glmnet_out,lambda==cv_out$lambda.1se)  # filters only to lambda which gives best CV
  # glmnet_fit_best <- filter(glmnet_out,lambda==cv_out$lambda.1se) # filters only to lambda which gives best CV + 1SE 
  glmnet_fit_best <- glmnet_fit_best %>% 
    select(seqid=term,estimate) %>% 
    arrange(desc(abs(estimate))) 
  
  glmnet_fit_best <- right_join(transl_5k[,-1],glmnet_fit_best)
  
  
  model_coefs_soma_ms[[j]] <- glmnet_fit_best
  
  selected_markers <- glmnet_fit_best$seqid
  
  coef_to_label <- paste("csf_ms_all",j,add,"corrplot.csv",sep="_")
  write_csv(glmnet_fit_best,paste("./output/EN_models_CEL/ms_all",j,add,"EN.csv",sep="_"))
  
  
  # save coeficients as unique object
  coefficients_pos_noNEF_soma <- glmnet_fit_best
  
  # save the model as RDS
  saveRDS(cv_out,paste0("./output/EN_models_CEL/EN_sin_",j,"_model_",add,"test.rds"))
  
  # plot predictions using validation data in all patients
  png(paste("./output/EN_models_CEL/EN_model_of",j,add,"observed_vs_predicted.png",sep="_"),width = 10, height = 5,res = 150, units = "in")
  par(mfrow=c(1,2))
  
  y_min <- .9*min(predict(cv_out$glmnet.fit,newx=x_all,s=cv_out$lambda.1se))
  y_max <- 1.1*max(predict(cv_out$glmnet.fit,newx=x_all,s=cv_out$lambda.1se))
  x_min <- .9*min(y_all,na.rm = TRUE)
  x_max <- 1.1*max(y_all,na.rm = TRUE)
  
  min_lim <- ifelse(x_min<y_min,x_min,y_min)
  max_lim <- ifelse(x_max>y_max,x_max,y_max)
  
  #training predictions
  predicted <- predict(cv_out$glmnet.fit,newx=x,s=cv_out$lambda.1se)
  pred <- as.numeric(predicted)
  obs <- as.numeric(y)
  
  rcoef <- cor.test(obs,pred,method="spearman")
  rsqr <- lm(pred~obs)
  concordance <- epi.ccc(pred,obs,ci = "z-transform",conf.level = .95)
  plot(pred~obs,xlab="observed",ylab="predicted",main=paste("EN soma for ",j,"\n",addd,"\nin TRAINING COHORT"),xlim=c(min_lim,max_lim),ylim=c(min_lim,max_lim))
  abline(rsqr,col="blue",lwd = 3)
  legend("topleft",legend=paste("Spearman Rho: ",round(rcoef$estimate,digits=3),
                                "  p-value: ", format.pval(rcoef$`p.value`,eps=0.001,digits=4),
                                "\nslope = ",round(rsqr$coefficients[2],digits = 3)," intercept = ",round(rsqr$coefficients[1],digits = 3),
                                "\nlm R2: ",round(summary(rsqr)$r.squared,digits = 3),sep=""),
         text.col = "blue",bty="n")
  abline(0,1,col="red",lwd = 2,lty =2)
  legend("bottomright",legend=paste("\nCCC = ",round(concordance$rho.c[1],digits=3)),text.col = "red",bty="n")
  
  train_intercept <- rsqr$coefficients[1]
  train_slope <- rsqr$coefficients[2]
  
  
  
  #save lm model for this variable 
  lm_models[[j]] <- rsqr$coefficients
  
  
  
  #validation predictions
  predicted <- predict(cv_out$glmnet.fit,newx=x_val,s=cv_out$lambda.1se)
  pred <- as.numeric(predicted)
  obs <- as.numeric(y_val)
  plot(pred~obs,xlab="observed",ylab="predicted",main=paste("EN model of",j,"\n",addd,"\nin VALIDATION COHORT"),xlim=c(min_lim,max_lim),ylim=c(min_lim,max_lim))
  rcoef <- cor.test(obs,pred,method="spearman")
  rsqr <- lm(pred~obs)
  concordance <- epi.ccc(pred,obs,ci = "z-transform",conf.level = .95)
  abline(rsqr,col="blue",lwd = 3)
  legend("topleft",legend=paste("Spearman Rho: ",round(rcoef$estimate,digits=3),
                                "  p-value: ", format.pval(rcoef$`p.value`,eps=0.001,digits=4),
                                "\nslope = ",round(rsqr$coefficients[2],digits = 3)," intercept = ",round(rsqr$coefficients[1],digits = 3),
                                "\nlm R2: ",round(summary(rsqr)$r.squared,digits = 3),sep=""),
         text.col = "blue",bty="n")
  abline(0,1,col="red",lwd = 2,lty =2)
  legend("bottomright",legend=paste("\nCCC = ",round(concordance$rho.c[1],digits=3)),text.col = "red",bty="n")
  
 

  dev.off()
   
 data <- data_all
  # save predictions for training and validation into soma_ms_en
  data[[paste("predicted",j,add,sep = "_")]] <- as.numeric(predict(cv_out$glmnet.fit,newx=data_out[,new_markers],s=cv_out$lambda.1se))
  
  # rename predicted outcome (it's the last column)
  names(data)[length(names(data))] <- "pred_cel"
  
  data_to_save <- data %>% 
    select(sampleid,patientcode,lpdate,diagnosis,age,cohort_split,all_of(marker),pred_cel)
  
  write.csv(data_to_save,"./output/EN_models_CEL/predicted_values_CEL_en_all_soma_only_withoutNEF_noPathogen_11062023.csv")
  



saveRDS(cv_list_soma_ms,paste("./output/EN_models_CEL/ms_models",add,".rds",sep="_"))
saveRDS(model_coefs_soma_ms,paste("./output/EN_models_CEL/ms_model_coefs",add,".rds",sep="_"))



# plot CEL predictions against age

p <- data_to_save %>% 
  filter(diagnosis %in% c("RR-MS","SP-MS","PP-MS","HD")) %>% 
  mutate(cohort=diagnosis) %>%
  mutate_at(.vars = "cohort",.funs = function(x){ifelse(x=="HD",yes = "HD",no = "MS")}) %>% 
  # mutate_at(.,.vars = "diagnosis",.funs = as.factor) %>% 
  ggplot(aes(x=age,y=pred_cel)) +
  geom_point(aes(fill=age,alpha=0.5,shape=cohort),color="darkgray",size=4,alpha=0.75)+
  scale_fill_viridis_c(option = "C")+
  scale_shape_manual(breaks = c("MS","HD"),values = c(21,23))+
  # xlab("msdss") +
  ylab("predicted CEL")+
  # scale_color_gradient(low = "yellow", mid="white",high = "red", na.value = NA)+
  
  # scale_fill_gradient(colours = viridis)+
  # scale_color_gradient2(low = 'yellow', mid = 'white', high = 'red')+
  theme_bw()+
  geom_smooth(aes(color=cohort),method = "lm", fill="lightgray",size=0.75)+
  theme(legend.background = element_rect(fill="transparent"),
        # legend.box.background = element_rect(colour = "black",fill="transparent"),
        legend.position = "right",
        legend.box.just = "left", #justification of the text in the legend
        legend.margin = margin(8, 8, 8, 8),
        legend.text=element_text(size=14),
        plot.margin = margin(0.5,.5,0.5,0.5,"cm"))
        # panel.border = element_rect(colour = "black", fill=NA, size=1),
        # panel.background = element_rect(fill = "transparent"), # bg of the panel
        # plot.background = element_rect(fill = "transparent", color = NA)) # bg of the plot)
        # legend.box.background = element_rect(fill = "transparent"))
p

ggsave(plot = p,device = "png",filename = "./output/EN_models_CEL/pred_cel_vs_age_all_soma_only_withoutNEF.png",width = 10,height = 7,units = "in")

###############################################################################
# identify which CEL surrogate values are above HD prediction interval
###############################################################################

# isolate MS and HD samples only 

data_to_plot <- data_to_save %>% 
  filter(diagnosis %in% c("RR-MS","SP-MS","PP-MS","HD")) %>% 
  mutate(cohort=diagnosis) %>%
  mutate_at(.vars = "cohort",.funs = function(x){ifelse(x=="HD",yes = "HD",no = "MS")})

# generate CEL model vs age in HD

hd_cel <- data_to_plot %>% 
  filter(diagnosis=="HD")

# build a model
hd_cel_mod <- lm(data = hd_cel,formula = pred_cel~age)
summary(hd_cel_mod)

# generate confidence and prediction interval lines 

# calculate confidence interval for CEL surrogate 
confidence_int <- predict(hd_cel_mod,data.frame(age=data_to_plot$age), interval="confidence",level=0.95)
#calculate prediction interval for CEL surrogate for HD
prediction_int <- predict(hd_cel_mod,data.frame(age=data_to_plot$age), interval="prediction",level=0.95)

# add upper prediction interval to the dataset 

data_to_plot$upper_prediction <- prediction_int[,3]

# generate a column identifying above and below prediction interval

data_to_plot$above_hd <- ifelse(test = data_to_plot$pred_cel>data_to_plot$upper_prediction,yes = "yes",no = "no")

#######!!!!!
# plot CEL predictions against age with confidence and prediction intervals 

p <- data_to_plot %>% 
  ggplot(aes(x=age,y=pred_cel)) +
  geom_point(aes(fill=age,alpha=0.5,shape=cohort,color=above_hd),size=4,alpha=0.75)+
  scale_fill_viridis_c(option = "C")+
  scale_color_manual(values = c("white","red"),breaks = c("no","yes"))+
  scale_shape_manual(breaks = c("MS","HD"),values = c(21,23))+
  # confidence
  geom_line(aes(y=confidence_int[,1]), color = "blue", linetype = "solid")+
  geom_line(aes(y=confidence_int[,2]), color = "purple", linetype = "dashed")+
  geom_line(aes(y=confidence_int[,3]), color = "purple", linetype = "dashed")+
  
  # prediction
  geom_line(aes(y=prediction_int[,2]), color = "red", linetype = "dashed")+
  geom_line(aes(y=prediction_int[,3]), color = "red", linetype = "dashed")+
  
  ylab("predicted CEL")+
  
  theme_bw()+
  # geom_smooth(aes(color=cohort),method = "lm", fill="lightgray",size=0.75)+
  theme(legend.background = element_rect(fill="transparent"),
        # legend.box.background = element_rect(colour = "black",fill="transparent"),
        legend.position = "right",
        legend.box.just = "left", #justification of the text in the legend
        legend.margin = margin(8, 8, 8, 8),
        legend.text=element_text(size=14),
        plot.margin = margin(0.5,.5,0.5,0.5,"cm"))
        # panel.border = element_rect(colour = "black", fill=NA, size=1),
        # panel.background = element_rect(fill = "transparent"), # bg of the panel
        # plot.background = element_rect(fill = "transparent", color = NA)) # bg of the plot)
        # legend.box.background = element_rect(fill = "transparent"))
p

ggsave(plot = p,device = "png",filename = "./output/EN_models_CEL/pred_cel_vs_age_all_soma_only_withoutNEF_wPredictionInt.png",width = 10,height = 7,units = "in")

to_save <- data_to_plot %>% 
  dplyr::select(sampleid,patientcode,lpdate,pred_cel,upper_prediction,above_hd)

write_csv(to_save,"./output/EN_models_CEL/cel_predicitions_wPrediction_interval_cutoff.csv")

# isolate samples with "above HD" CEL predictions
above_hd_selection <- data_to_plot %>% 
  filter(above_hd=="yes")

# plot NFL vs predicted CEL in samples above HD only

nfl_to_plot <- data %>% 
  filter(diagnosis %in% c("RR-MS","SP-MS","PP-MS") & !is.na(nfl_csf_age)) %>% 
  mutate(NFL=log10(nfl_csf_age+1000)) %>% 
  filter(sampleid %in% above_hd_selection$sampleid)

# generate NFL-predicted CEL residuals

nfl_to_plot$NFL_CEL_residuals <- summary(lm(nfl_to_plot$NFL~nfl_to_plot$pred_cel))$residuals

# save as csv file
write_csv(nfl_to_plot,"./output/EN_models_CEL/nfl_cel_residuals.csv")

# correlate NFL-CEL residuals with somascan
suppressWarnings(
nfl_cel_res_cor <- corr_fun(dat = nfl_to_plot,outcomes = "NFL_CEL_residuals",high_markers = high_markers)
)
names(nfl_cel_res_cor)[1] <- "seqid"

nfl_cel_res_cor <- right_join(transl_5k[,-1],nfl_cel_res_cor)

# save as CSV

write_csv(nfl_cel_res_cor,"./output/EN_models_CEL/NFL_CEL_residuals_corr_wSoma.csv")

# plot NFL vs predicted CEL 
p <- nfl_to_plot %>% 
  ggplot(aes(y=NFL,x=pred_cel)) +
  # geom_rect(aes(xmin=-Inf,xmax=Inf,ymin=hd_mean-hd_sd,ymax=hd_mean+hd_sd),fill="lightgreen")+
  geom_point(shape=21,color="darkgray",size=2,alpha=0.75)+
  # xlim(c(-0.3,2.3))+
  ylab("log10(HD-adj CSF NFL + 1000)") +
  xlab("predicted CEL")+
  # scale_color_gradient(low = "yellow", mid="white",high = "red", na.value = NA)+
  
  # scale_fill_gradient(colours = viridis)+
  # scale_color_gradient2(low = 'yellow', mid = 'white', high = 'red')+
  theme_bw()+
  geom_smooth(method = "lm", fill="lightgray",size=0.75)+
   stat_regline_equation( aes(label =  paste(..eq.label.., ..adj.rr.label.., sep = "~~~~")),formula = y~x)+
  theme(legend.background = element_rect(fill="transparent"),
        # legend.box.background = element_rect(colour = "black",fill="transparent"),
        legend.position = "right",
        legend.box.just = "left", #justification of the text in the legend
        legend.margin = margin(8, 8, 8, 8),
        legend.text=element_text(size=14),
        plot.margin = margin(0.5,.5,0.5,0.5,"cm"))
        # panel.border = element_rect(colour = "black", fill=NA, size=1),
        # panel.background = element_rect(fill = "transparent"), # bg of the panel
        # plot.background = element_rect(fill = "transparent", color = NA)) # bg of the plot)
        # legend.box.background = element_rect(fill = "transparent"))
p

ggsave(plot = p,device = "png",filename = "./output/EN_models_CEL/pred_cel_vs_NFL.png",width = 5,height = 3,units = "in")

```
